requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
redis==5.0.8
kiteconnect==5.0.1
jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pytz>=2024.1
itsdangerous>=2.2.0
pydantic[email]>=2.6
cryptography>=41.0.0
python-dotenv>=1.0.0
slowapi